+++ tests/Core/utils/PriorityConvertorTest.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
-#include "Core/persistence/PriorityConvertor.h"
+#include "Core/utils/PriorityConvertor.h"
 
 class PriorityConvertorTest : public ::testing::Test {
